Implemented mini2Dx timestep logic in the UAT
Separated the UAT from the mini2Dx initialization logic
